refactor: changed public access to package-private access
